feat: check collection lock (#289)
* feat: verify collection locks for ethereum wearables

* chore: unnecessary restore

* feat: feedback and better tests

* feat: split publish and lock checks in two methods

* feat: possitve lock test, correct types, lock status code

* feat: correct naming and checks for item router

+++ src/Collection/Collection.router.ts
@@ -15,18 +15,23 @@ import { peerAPI, Wearable } from '../ethereum/api/peer'
 import { FactoryCollection } from '../ethereum'
 import { Ownable } from '../Ownable'
 import { Item, ItemAttributes } from '../Item'
-import { Collection, CollectionAttributes } from '../Collection'
+import {
+  Collection,
+  CollectionService,
+  CollectionAttributes,
+} from '../Collection'
 import { isCommitteeMember } from '../Committee'
 import { collectionSchema, saveTOSSchema } from './Collection.types'
 import { RequestParameters } from '../RequestParameters'
 import { hasAccess } from './access'
-import { isPublished } from '../utils/eth'
 import { ItemFragment } from '../ethereum/api/fragments'
 import { sendDataToWarehouse } from '../warehouse'
 
 const validator = getValidator()
 
 export class CollectionRouter extends Router {
+  public service = new CollectionService()
+
   mount() {
     const withCollectionExists = withModelExists(Collection, 'id')
     const withCollectionAuthorization = withModelAuthorization(Collection)
@@ -56,7 +61,6 @@ export class CollectionRouter extends Router {
      */
     this.router.get(
       '/collections/:id',
-      withAuthentication,
       withCollectionExists,
       server.handleRequest(this.getCollection)
     )
@@ -291,15 +295,13 @@ export class CollectionRouter extends Router {
     const eth_address = req.auth.ethAddress
 
     try {
-      await Collection.update(
-        { lock: new Date(Date.now()) },
-        { id, eth_address }
-      )
-      return true
+      const lock = new Date(Date.now())
+      await Collection.update({ lock }, { id, eth_address })
+      return lock
     } catch (error) {
       throw new HTTPError(
         "The collection couldn't be updated",
-        { id, eth_address, error },
+        { id, eth_address, error: (error as Error).message },
         STATUS_CODES.error
       )
     }
@@ -348,12 +350,24 @@ export class CollectionRouter extends Router {
       )
     }
 
-    if (await this.isCollectionPublished(id)) {
-      throw new HTTPError(
-        "The collection is published. It can't be updated",
-        { id },
-        STATUS_CODES.unauthorized
-      )
+    const collection = await Collection.findOne<CollectionAttributes>(id)
+
+    if (collection) {
+      if (await this.service.isPublished(collection.contract_address)) {
+        throw new HTTPError(
+          "The collection is published. It can't be saved",
+          { id },
+          STATUS_CODES.unauthorized
+        )
+      }
+
+      if (this.service.isLockActive(collection.lock)) {
+        throw new HTTPError(
+          "The collection is locked. It can't be saved",
+          { id },
+          STATUS_CODES.locked
+        )
+      }
     }
 
     if (id !== attributes.id) {
@@ -376,7 +390,9 @@ export class CollectionRouter extends Router {
   deleteCollection = async (req: AuthRequest) => {
     const id = server.extractFromReq(req, 'id')
 
-    if (await this.isCollectionPublished(id)) {
+    const collection = (await Collection.findOne(id)) as CollectionAttributes // existance checked on middleware
+
+    if (await this.service.isPublished(collection.contract_address)) {
       throw new HTTPError(
         "The collection is published. It can't be deleted",
         { id },
@@ -384,34 +400,18 @@ export class CollectionRouter extends Router {
       )
     }
 
+    if (this.service.isLockActive(collection.lock)) {
+      throw new HTTPError(
+        "The collection is locked. It can't be deleted",
+        { id },
+        STATUS_CODES.locked
+      )
+    }
+
     await Promise.all([
       Collection.delete({ id }),
       Item.delete({ collection_id: id }),
     ])
-    return true
-  }
-
-  async isCollectionPublished(collectionId: string) {
-    const dbCollection = await Collection.findOne<CollectionAttributes>(
-      collectionId
-    )
-
-    if (!dbCollection) {
-      return false
-    }
-
-    const remoteCollection = await collectionAPI.fetchCollection(
-      dbCollection.contract_address
-    )
-
-    // Fallback: check against the blockchain, in case the subgraph is lagging
-    if (!remoteCollection) {
-      const isCollectionPublished = await isPublished(
-        dbCollection.contract_address
-      )
-      return isCollectionPublished
-    }
-
     return true
   }
 }
